Separate Service and Service Interfaces
