+++ src/main/java/edu/warbot/config/DataSourceConfig.java
@@ -78,7 +78,7 @@ public class DataSourceConfig
     public JpaVendorAdapter jpaVendorAdapter() {
         HibernateJpaVendorAdapter hibernateJpaVendorAdapter = new HibernateJpaVendorAdapter();
         hibernateJpaVendorAdapter.setShowSql(showSQL);
-        hibernateJpaVendorAdapter.setGenerateDdl(true);
+        hibernateJpaVendorAdapter.setGenerateDdl(false);
         return hibernateJpaVendorAdapter;
     }
 
@@ -86,7 +86,6 @@ public class DataSourceConfig
     @Bean
     public PlatformTransactionManager transactionManager(
             EntityManagerFactory entityManagerFactory) {
-
         return new JpaTransactionManager(entityManagerFactory);
     }
