+++ src/main/java/edu/warbot/config/WebSocketConfig.java
@@ -35,7 +35,7 @@ public class WebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer {
     @Bean
     public CodeEditorListener codeEditorListener(SimpMessagingTemplate messagingTemplate) {
         CodeEditorListener codeEditorListener = new CodeEditorListener(messagingTemplate);
-        codeEditorListener.setLogIn("/editor/register");
+        codeEditorListener.setAccess("/editor/register");
         return codeEditorListener;
     }
 }
