+++ src/main/java/edu/warbot/controllers/AccountController.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
 import org.springframework.util.Assert;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,13 +23,12 @@ class AccountController {
     @Autowired
     private AccountRepository accountRepository;
 
-    @RequestMapping(value = "account/current", method = RequestMethod.GET)
-    @ResponseStatus(value = HttpStatus.OK)
-    @ResponseBody
-    public Account accounts(Principal principal) {
+    @Transactional
+    private Account accounts(Principal principal) {
         Assert.notNull(principal);
         return accountRepository.findByEmail(principal.getName());
     }
+
     @RequestMapping(value = "account/profile", method = RequestMethod.GET)
     public String profile(Model model,Principal principal)
     {
