+++ src/main/java/edu/warbot/controllers/WebGameController.java
@@ -5,7 +5,7 @@ import edu.warbot.models.Party;
 import edu.warbot.online.messaging.WebGameSettings;
 import edu.warbot.repository.AccountRepository;
 import edu.warbot.services.WarbotOnlineService;
-import edu.warbot.services.WebGameService;
+import edu.warbot.services.WebGameServiceImpl;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,7 +26,7 @@ public class WebGameController {
     private static final Log logger = LogFactory.getLog(WebGameController.class);
 
     @Autowired
-    private WebGameService webGameService;
+    private WebGameServiceImpl webGameService;
 
     @Autowired
     private AccountRepository accountRepository;
