+++ src/main/resources/db/ddl_hsql.sql
@@ -2,16 +2,16 @@
     create table ACCOUNT (
         id bigint generated by default as identity (start with 1),
         account_email varchar(255),
-        account_firstname varchar(255) not null,
+        account_firstName varchar(255) not null,
         account_inscriptionDate timestamp,
         account_activated boolean not null,
         account_premium boolean not null,
         account_lastConnectionDate timestamp,
-        account_lastname varchar(255) not null,
+        account_lastName varchar(255) not null,
         account_password varchar(255) not null,
         account_expirationDate timestamp,
         account_role varchar(255),
-        account_screenname varchar(255) not null,
+        account_screenName varchar(255) not null,
         primary key (id)
     );
 
@@ -70,7 +70,7 @@
         add constraint UK_jkd9cix86ciqqcm6do9odr3h7  unique (account_email);
 
     alter table ACCOUNT 
-        add constraint UK_f4lc6mbyr2oahgta5npnx74l  unique (account_screenname);
+        add constraint UK_hnequcyk3hh0s97ligk9ijody  unique (account_screenName);
 
     alter table PARTY 
         add constraint UK_6f9crrkedyvu40ib31ca4s3w3  unique (party_name);
