+++ src/main/webapp/WEB-INF/views/configuration-editor/create.html
