+++ src/main/webapp/WEB-INF/views/configuration-editor/select.html
